Cast inserted primary key to int to avoid types conflicts at update in where clause.
git-svn-id: https://subversion.assembla.com/svn/file_sender/filesender/branches/filesender-2.0@2873 8555b757-d854-4b86-925a-82fd84c90ff4

+++ classes/data/File.class.php
@@ -125,7 +125,7 @@ class File extends DBObject
             $this->updateRecord($this->toDBData(), 'id');
         }else{
             $this->insertRecord($this->toDBData());
-            $this->id = DBI::lastInsertId();
+            $this->id = (int)DBI::lastInsertId();
         }
     }
